backend/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn==1.6.1
joblib
tensorflow==2.19.0
keras==3.13.2
pillow
python-multipart
supabase
python-dotenv